Import illink.targets for ref projects (#118144)
This lets the logic from the ILLink targets add
`AssemblyMetadata("IsTrimmable", "true")` to ref projects which have
trimming enabled.

Also moves the trimming opt-out into `Directory.Build.props` when it
should be shared between src and corresponding ref projects.

This will be used to enable the Roslyn analyzer to produce a warning
for references to assemblies that aren't marked "IsTrimmable".