JIT: fix issue in conditional escape analysis
Conditional escape analysis is triggered by a GDV guarded call to
`GetEnumerator`. On the fast path this call is devirtualized and inlined
and exposes one or more enumerator object allocations. This allows
for collections to do special-case optimizations like returning a static
enumerator object for an empty collection.

When there is more than one enumerator object allocation under the GDV guard,
ensure that only the allocation that was analyzed for conditional escape is stack
allocated; the remaining enumerators will end up on the "slow path" where enumeration
happens via interface calls, and so must be heap objects.

Fixes #111922.