Deduplicate base interfaces from another assembly in COM generator (#111931)
Fixes cross-assembly interface inheritance.

Deduplicate external ComInterfaceInfos in the pipeline.

Don't remove external interface methods in the generator pipeline - these are required to create the new VTable struct type introduced in #116289.

Creates a new SourceAvailableIncrementalMethodStubGenerationContext for methods with source available and use IncrementalMethodStubGenerationContext as a base for externally defined methods.

Moves the ComInterfaces from SharedTypes to the Common folder, and defines a limited set of base interfaces in SharedTypes for interfaces in ComInterfaceGenerator to derived from. Adds a number of tests for different levels of inheritance across assemblies.