Simplify pragma suppressions in generated regex code (#118258)
Every generated regex output contained multiple pragma suppressions for warnings we know to be possible. We can simplify this by just suppressing all warnings in the generated code. If there is a stray warning (e.g. because of a custom analyzer), there's nothing the developer can do about it, and it can break their build if they have warnings-as-errors.