Implement ReadSpan/WriteSpan methods in UnmanagedMemoryAccessor
The current UnmanagedMemoryAccessor only has the ability to read/write arrays using a slow struct marshalling loop, despite the fact that the underlying SafeBuffer now has a much more performant ReadSpan/WriteSpan implementation. This change adds a shim method to UnmanagedMemoryAccessor which allows it to read/write byte span data with a similar signature to Read/WriteArray and the performance of Span.