Add trim/aot analyzer warnings for transitive references (#118180)
Adds new analyzer warnings for references to assemblies not marked as
trim or aot compatible.

The new warnings are opt-in under the following settings:

- `<VerifyReferenceTrimCompatibility>true</VerifyReferenceTrimCompatibility>`
- `<VerifyReferenceAotCompatibility>true</VerifyReferenceAotCompatibility>`

The analyzer uses AssemblyMetadataAttribute in the referenced
assemblies to detect whether they are marked as trim/aot
compatible. To mark a project as trim/aot compatible, it should be
built with the usual recommended property settings:

- `<IsTrimmable>true</IsTrimmable>`
  adds `[assembly: AssemblyMetadata("IsTrimmable", "True")]`
- `<IsAotCompatible>true</IsAotCompatible>`
  adds `[assembly: AssemblyMetadata("IsAotCompatible", "True")`