Use (Un)ManagedValuesSource.Length as length condition in marshalling and unmarshalling loops (#118190)
Instead of assuming the Span returned from Get(Un)ManagedValuesSource is the same length as the length parameter, use the length of the span returned from the marshaller methods in the for loop condition. This allows the marshaller to recognize native values that have a different length than expected. In particular, if an array parameter is null, allow the ArrayMarshaller to return empty spans.

Also, updates the Array, Span, ReadOnlySpan, and PointerArray marshallers to check for null unmanaged values when creating unmanaged source/destination spans.

Co-authored-by: Aaron Robinson <arobins@microsoft.com>