Remove 2GB allocations from MemoryStream_SeekOverflow_Throws.
This test was allocating arrays that were nearly 2GB in size in order to attempt to test an edge case around seek overflowing. These large allocations are very unable in CI and started failing for Android + CoreCLR.

The test is already disabled for Linux, for iOS, for tvOS, and partially for Windows. Rather than disable it yet again, let's just not test the very large allocation sizes. At this point the test has been disabled on so many platforms for being problematic we aren't getting much value from the large allocations.