Handle 'lib' variants in ModuleNameVariations to generate DirectPInvokes (#118285)
In ILC, when a DirectPInvokes item was prefixed with 'lib', it wouldn't recognize the library and wouldn't generate a direct pinvoke. There are a lot of places within the repo (and I assume other codebases) that use the 'lib' prefix in DirectPInvokes, so it seems to make more sense to handle that case rather than remove the prefix from all usages.

This accounts for the scenario with `[DllImport("libMyLibrary")]` and `--directpinvoke MyLibrary`, but not the scenario with `[DllImport("MyLibrary")]` and `--directpinvoke:libMyLibrary`.

Co-authored-by: Jan Kotas <jkotas@microsoft.com>