JIT: Build community jits on Windows in superpmi-diffs
When we changed to run superpmi-diffs on macOS arm64, we lost the build
of the community JITs on Windows. This was switched over to build in
superpmi-replay, but that pipeline does not run on all JIT PRs.

Readd the Windows build back to ensure compiler errors do not slip in
for the community JITs.