Fix rebuilding mono aot cross compiler when building locally
This is a side-effect of the changes made in https://github.com/dotnet/runtime/pull/109612.
When we do rerun the offsets generation script locally it causes the file to be touched even if it's identical, which in turn causes the whole aot cross compiler binary to be rebuilt.

To fix this only write to the output offsets file when it actually changed.