JIT: create inferred GDVs for enumerator var uses that lack PGO
Long-running enumerator loops at Tier0+instr will see their executions
transition over to a non-instrumented OSR version of the code.

This can cause loss of PGO data in the portions of the method that
execute after the leaving the loop that inspires OSR.

For enumerator vars we can safely deduce the likely classes from
probes made earlier in the method. So when we see a class profile
for an enumerator var, remember it and use it for subsequent calls
that lack their own profile data.

Addresses part of #118420.