Implement type name resolution for ILLink analyzer
Adds Roslyn analyzer support for parsing type names passed to
Type.GetType.

Reapplies e5b1d021938d41b589feeb5b6e9dcb55035dfda9 which was reverted
due to https://github.com/dotnet/runtime/issues/106321. Now that
Roslyn has updated to reference System.Reflection.Metadata 9.0, we
could use `TypeName` from that assembly. However, the code was using
`TypeNameHelpers.Unescape` which was moved to `TypeName.Unescape` in
.NET 10.

Instead of using the 9.0.0 version, this just source includes the
sources for the TypeNameResolver to ensure we get the same behavior in
the analyzer as we have in ILC/ILLink.