Fix unhandled exception reporting corner cases (#118450)
* Fix unhandled exception reporting corner cases

There are several issues with processing exceptions that are not handled
by managed code on Windows:
* Sometimes the AppDomain.UnhandledException event is sent multiple
  times
* Exception flowing to foreign native code is reported as unhandled even
  when it is actually caught by the native code
* In rare cases, the unhandled exception stack trace is doubled

This change fixes them by not reporting the unhandled exception in the
SfiNext on Windows. It just raises the underlying SEH exception there
with the thread marked so that the personality routines for the managed
frames won't run the managed exception handling code. If the exception
is truly unhandled, the `InternalUnhandledExceptionFilter_Worker` will be
called by the unhandled exception filter installed for the process and
report the exception as unhandled.
If that exception ends up being caught by a foreign native code, then
nothing will be reported.

Close #115215

* Ensure AppDomain.CurrentDomain.UnhandledException occurs before finally