Add _UseNativeLibPrefix option for publishing a NativeAOT shared library on Unix (#118299)
Most Unix systems expect the library filename to be libMyLibrary.so / libMyLibrary.dylib. On Java on android, libraries will not be found unless they follow this naming format.

Tring to work around this by renaming the TargetName or AssemblyName can cause rd.xml files to fail to resolve the assembly in ilc, and it doesn't feel right to .

This PR adds a _UseNativeLibPrefix option to add the "lib" prefix to the output library when it's a shared library published for a non-windows OS. It's opt-in-only to avoid breaking existing setups, but it might be worth considering as a default in future releases to match the rest of the unix ecosystem.