Add NativeAOT on Android test (#118037)
Adds a project to run a NativeAOT shared library on Android.

The android test builds a single native library that exports the methods that are expected by the monodroid tests (InitRuntime, SetEnv, ExecEntryPoint, FreeUnmanagedResources). JNI struct definitions are copied to get arg strings from the JNIEnv object.

ApkBuilder no longer needs to compile the monodroid native library, but it does need to add the .jar from the cryptography library to the apk bundle in order to run crypto operations.

Build targets were updated to account for the new NativeAOT tests, and a new job was added for NativeAOT on android tests.

Co-authored-by: Filip Navara <filip.navara@gmail.com>