Trim custom attributes
1. To be able to read custom attributes at runtime, we need to use the internal metadata APIs
2. When the internal metadata API is not used, we know nobody could be reading the attributes
3. Make custom attribute emission conditional on the presence of the reading API in the graph
4. Use separate conditions for different kinds of attributes - fields, types, methods, etc.

In practice, this doesn't help much outside of Hello World scenarios because custom attribute reading at runtime happens through a virtual method on `MemberInfo`. So if e.g. `PropertyInfo` is allocated and somebody calls `Type.GetCustomAttributes` (which is `MemberInfo.GetCustomAttributes` virtual), we include code to read attributes on properties too. This is a solvable problem if we learn to do devirtualization during scanning phase; then we can reap more benefit from this.