Fix unloadability crash
There is a rare case when attempt to unload an assembly load context in
multi-assembly load context scenario leads to access violation in the
runtime.

The crash is caused by a DomainAssembly pointer in the AppDomain
assembly list not being removed while the DomainAssembly was deleted.

There is a detailed analysis in the #116953.

This change fixes it and adds a regression test that is a repro app
provided in the issue converted to coreclr test.

Close #116953