Only add GVMDependencies for types with used GVMs (#118704)
Ever since GVM support was added to native AOT, we were generating the GVM resolution metadata for every type considered allocated. This included GVMs that were never even called: see `TypeGVMEntriesNode` that simply goes over everything on the type - we were adding a `TypeGVMEntriesNode` for all allocated types that have something to do with GVMs.

This PR changes it so that we only generate `TypeGVMEntries` for types that have at least one used GVM. This is a scoped down version of https://github.com/dotnet/runtime/pull/118632 (that tries to track things per-method) with a lot less risk.