Correctly lock around iterating the AvailableParamTypes hashtable when adding fields/classes via ApplyMetadata (#118608)
This does not eliminate all possible race conditions here, if a type is in progress loading when the metadata is updated, then there is still a race, but this should narrow the possible races in this space. Additionally, this should address the debug asserts which have been hit in this space.

Fixes #115318