Fix check for `host_runtime_contract` having `external_assembly_probe` (#118792)
We should check the size of the contract to determine if it can have `external_assembly_probe`. In practice, we probably haven't hit this since `hostpolicy` ships next to `coreclr` - but any sort of patching such that `coreclr` is newer than `hostpolicy` could result in incorrectly determining that `external_assembly_probe` is set.