Serialize negative constants according to the operand type
Prior to this patch, any negative constants smaller than -0xFFFFFF would
get serialized to 64-bit constants, so we would end up with assembly
statements like `imul eax, esi, 0xFFFFFFFFFEFFFFFF`, which is invalid
since the constant is larger than 32 bits, which is the size of `eax`
and `esi` registers.

This patch fixes the code so that for known 1-, 2, 4-, and 8-byte types,
we first truncate the constant to the correct number of bits before
printing it.

Fix #118813