Fix deadlock when creating threads from ModuleInitializer in a shared library (#118928)
The problem is that the first time we reverse p/invoke into an API in the shared library, we need to do two things: initialize runtime and run module initializers. Running module initializers is done as part of runtime initialization in the shared library case because it was convenient to do it there (in the EXE case, we run them after runtime initialization, right before Main).

The problem with running it as part of the runtime initialization is that the arbitrary user code can deadlock us (see the issues for stack).

This fixes the problem by attaching the newly created thread to the runtime before we do the reverse p/invoke. Since thread is already attached, we don't wait for runtime initialization. Waiting is unnecessary because the runtime needs to be really far along the initialization path if we got to a managed `Thread.Start` call.

Fixes #118773. Fixes #107699.