Fix GCStress regression (#119028)
* Fix GCStress regression

GCStress is not able to find the original version of the code for PInvoke stubs after #117901. We need to compensate
for the MethodDesc adjustment done for PInvoke stubs when storing the original version of the code during GC stress.

* Partial revert of #117901 for vararg stubs

Interop marshalling of varargs needs MethodDesc calling convention to support ldftn <PInvoke method with varargs>. It is not possible to smuggle the target MethodDesc* via vararg cookie in this case.

Co-authored-by: Aaron Robinson <arobins@microsoft.com>