Move ILLink generic handling to scanner (#118708)
This moves the ILLink generic dataflow analysis out of MarkStep and
into the scanner, unifying it with the logic (not actually sharing
code) in ILC and ILLink.RoslynAnalyzer. It is essentially a port of
the ILC logic into ILLink.

This ran into an issue with incorrect suppressions in DI (see
https://github.com/dotnet/runtime/issues/81358) that we have a
workaround for in native AOT, so this implements a similar workaround
in ILLink. This includes a copy of the native AOT smoke test for that
workaround, adjusted to run as a shared testcase.

Fixes https://github.com/dotnet/runtime/issues/113249