Use fast hashing helpers for types with trailing padding (#118986)
This is same as #111620 but with an extra thing: enable unconditional method body folding on `__GetFieldHelper` overrides (both the simplified ones that just return a negative number and normal ones).

Before this PR we considered valuetypes as byte-hashable/comparable only if the valuetype didn't have any trailing padding.

This fixes things up so that we can consider types with trailing padding byte-hashable/comparable if the other requirements still hold (such as no double/float fields, no GC references, etc.) and there is a trailing padding. In this case, we remember the number of bytes that we should be looking at and ignore the rest.

Since this doesn't seem to provide meaningful size savings according to rt-sz, this is a throughput improvement, not a size improvement.