JIT: Disallow widening tailcalls in async functions
The JIT allows tailcalls that depend on widening happening in the
callee: e.g. an int32 returning function can tailcall an int16 returning
function because the managed ABI specifies that the callee will widen
anyway.

However, this is not legal for async methods. For async methods the
int32 and int16 returning functions come with two different resumption
stubs that are responsible for propagating the returned result into the
next continuation in asynchronous cases. Allowing an int32 -> int16
tailcall means only a 16-bit value gets propagated into a caller that
expects 32 bits to be written into its continuation.

Fix #119060