JIT: fix crash in LSRA seen on VMR build on AVX-512 machines
During the second stage bootstrap build VMR on an AVX-512 capable machine,
we end up in `try_SPILL_COST` looking at a K-reg spill candidate without an
assigned interval,  and crash.

The code was already handling this sort of thing for ARM64; so we now extend
it to cover all architectures.

Fixes the crash seen in #119070.