Fix missing cleanup of cached last used code heap
I've started to hit a problem with the interpreter when a CodeHeap
pointer obtained from the m_pLastUsedInterpreterDynamicCodeHeap
was invalid.
This issue is that we were missing cleanup of that cached code heap
pointer when the code heap was destroyed.

This change adds the cleanup in the same way it is done for
non-interpreter dynamic code heap.